+++ FunctionSpace/HexNodeBasis.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-HexNodeBasis::HexNodeBasis(int order){
+HexNodeBasis::HexNodeBasis(size_t order){
   /*
   // Set Basis Type //
   this->order = order;
@@ -36,10 +36,10 @@ HexNodeBasis::HexNodeBasis(int order){
   const int edgeV[2][12][2] =
     {
       { {0, 1}, {0, 3}, {0, 4}, {1, 2}, {1, 5}, {2, 3},
-	{2, 6},	{3, 7}, {4, 5}, {4, 7}, {5, 6}, {6, 7} },
+        {2, 6}, {3, 7}, {4, 5}, {4, 7}, {5, 6}, {6, 7} },
 
       { {1, 0}, {3, 0}, {4, 0}, {2, 1}, {5, 1}, {3, 2},
-	{6, 2},	{7, 3}, {5, 4}, {7, 4}, {6, 5}, {7, 6} },
+        {6, 2}, {7, 3}, {5, 4}, {7, 4}, {6, 5}, {7, 6} },
     };
 
   const int faceV[6][4] =
@@ -102,43 +102,43 @@ HexNodeBasis::HexNodeBasis(int order){
   // Vertex Based (Lagrange) //
   (*basis)[0] =
     new Polynomial((Polynomial(1, 0, 0, 0) - Polynomial(1, 1, 0, 0)) *
-		   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 1, 0)) *
-		   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 0, 1)));
+                   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 1, 0)) *
+                   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 0, 1)));
 
   (*basis)[1] =
     new Polynomial((Polynomial(1, 1, 0, 0))                          *
-		   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 1, 0)) *
-		   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 0, 1)));
+                   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 1, 0)) *
+                   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 0, 1)));
 
   (*basis)[2] =
     new Polynomial((Polynomial(1, 1, 0, 0)) *
-		   (Polynomial(1, 0, 1, 0)) *
-		   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 0, 1)));
+                   (Polynomial(1, 0, 1, 0)) *
+                   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 0, 1)));
 
   (*basis)[3] =
     new Polynomial((Polynomial(1, 0, 0, 0) - Polynomial(1, 1, 0, 0)) *
-		   (Polynomial(1, 0, 1, 0))                          *
-		   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 0, 1)));
+                   (Polynomial(1, 0, 1, 0))                          *
+                   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 0, 1)));
 
   (*basis)[4] =
     new Polynomial((Polynomial(1, 0, 0, 0) - Polynomial(1, 1, 0, 0)) *
-		   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 1, 0)) *
-		   Polynomial(1, 0, 0, 1));
+                   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 1, 0)) *
+                    Polynomial(1, 0, 0, 1));
 
   (*basis)[5] =
     new Polynomial((Polynomial(1, 1, 0, 0))                          *
-		   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 1, 0)) *
-		   Polynomial(1, 0, 0, 1));
+                   (Polynomial(1, 0, 0, 0) - Polynomial(1, 0, 1, 0)) *
+                    Polynomial(1, 0, 0, 1));
 
   (*basis)[6] =
     new Polynomial((Polynomial(1, 1, 0, 0)) *
-		   (Polynomial(1, 0, 1, 0)) *
-		   Polynomial(1, 0, 0, 1));
+                   (Polynomial(1, 0, 1, 0)) *
+                    Polynomial(1, 0, 0, 1));
 
   (*basis)[7] =
     new Polynomial((Polynomial(1, 0, 0, 0) - Polynomial(1, 1, 0, 0)) *
-		   (Polynomial(1, 0, 1, 0))                          *
-		   Polynomial(1, 0, 0, 1));
+                   (Polynomial(1, 0, 1, 0))                          *
+                    Polynomial(1, 0, 0, 1));
 
 
   // Edge Based //
@@ -152,8 +152,8 @@ HexNodeBasis::HexNodeBasis(int order){
   for(int l = 1; l < order; l++){
     for(int e = 0; e < 12; e++){
       (*basis)[i] = new Polynomial(
-	legendre[l].compose(lifting[edge1[e]] - lifting[edge2[e]]) *
-	(*(*basis)[edge1[e]] + *(*basis)[edge2[e]]));
+        legendre[l].compose(lifting[edge1[e]] - lifting[edge2[e]]) *
+        (*(*basis)[edge1[e]] + *(*basis)[edge2[e]]));
 
       i++;
     }
@@ -188,12 +188,12 @@ HexNodeBasis::HexNodeBasis(int order){
   for(int l1 = 1; l1 < order; l1++){
     for(int l2 = 1; l2 < order; l2++){
       for(int f = 0; f < 6; f++){
-	(*basis)[i] = new Polynomial(
-	  legendre[l1].compose(xi[f])  *
-	  legendre[l2].compose(eta[f]) *
-	  lambda[f]);
+        (*basis)[i] = new Polynomial(
+          legendre[l1].compose(xi[f])  *
+          legendre[l2].compose(eta[f]) *
+          lambda[f]);
 
-	i++;
+        i++;
       }
     }
   }
@@ -211,12 +211,12 @@ HexNodeBasis::HexNodeBasis(int order){
   for(int l1 = 1; l1 < order; l1++){
     for(int l2 = 1; l2 < order; l2++){
       for(int l3 = 1; l3 < order; l3++){
-	(*basis)[i] =
-	  new Polynomial(legendre[l1].compose(px) *
-			 legendre[l2].compose(py) *
-			 legendre[l3].compose(pz));
+        (*basis)[i] =
+          new Polynomial(legendre[l1].compose(px) *
+                         legendre[l2].compose(py) *
+                         legendre[l3].compose(pz));
 
-	i++;
+        i++;
       }
     }
   }
